configure store moved to proper location.
